% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoo.accumulate.data.R
\name{hoo.accumulate.data}
\alias{hoo.accumulate.data}
\title{Create rENA Accumulated Model based on Horizon of Observation Algorithm}
\usage{
hoo.accumulate.data(data, Units, Conversation, Codes, dataModeCol,
  modeObserve, usersCol, windowSize, ...)
}
\arguments{
\item{data}{The original data containing the multimodal data (input a data frame or a data table)}

\item{Units}{ENA Units used on the original data (need to input a character string or a vector of character strings)}

\item{Conversation}{ENA Conversations within the original data (need to input a character string or a vector of character strings)}

\item{Codes}{Big C Codes used in the original data (need to input a character string or a vector of character strings)}

\item{dataModeCol}{Name of the column where the types of multimodal data is stored (need to input a character string)}

\item{modeObserve}{Name of the mode of data where other players can observe their actions (need to input a character string or a vector of character strings)}

\item{usersCol}{Name of the column where usernames or userid (something to distinguish players in the game) is stored (need to input a character string)}

\item{windowSize}{Size of the moving stanza window, for looking backwards (need to input a numeric value; for whole conversation, input 1)}

\item{...}{Additional arguments to pass onto ena.accumulate.data function}
}
\value{
a data frame containing the adjacency vectors of each ENA Units within your data
}
\description{
For multimodal data, this function creates the correct rENA accumulated model that can be used to create rENA sets for analysis.
}
\examples{
accum = hoo.accumulate.data(data = mock, 
                            Units = c("site", "userName"), 
                            Conversation = c("site"), 
                            Codes = c("Code1", "Code2", "Code3", "Code4"), 
                            dataModeCol = "data", 
                            modeObserve = "chat", 
                            usersCol = "userName", 
                            windowSize = 4)

}
