% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoo.mc.ena.accumulate.data.R
\name{hoo.mc.ena.accumulate.data}
\alias{hoo.mc.ena.accumulate.data}
\title{Create rENA Accumulated Model based on Horizon of Observation Algorithm, Utilizing Multicore Computation}
\usage{
hoo.mc.ena.accumulate.data(data, Units, Conversation, Codes, dataModeCol,
  modeObserve, usersCol, windowSize, ...)
}
\arguments{
\item{data}{Multimodal data.frame or data.table}

\item{Units}{A vector of Strings describing the Units for ENA model}

\item{Conversation}{A vector of Strings describing the Conversations took place in ENA model}

\item{Codes}{A vector of Strings entailing the big C Codes of interest for ENA 
analysis}

\item{dataModeCol}{Name of the column where the types of multimodal data is stored}

\item{modeObserve}{Modes of data where actions are observable to all players}

\item{usersCol}{Name of the column entailing the unique user tracking info}

\item{windowSize}{Size of the moving stanza window, for looking backwards (for whole 
conversation, input 1)}

\item{...}{Additional arguments to pass along to rENA ena.accumulate.data method}
}
\value{
a data frame containing the adjacency vectors of each ENA Units within data
}
\description{
For multimodal data, this function creates the correct rENA accumulated model that can be used 
to create rENA sets for analysis. Specifically, this function utilizes parallel package's 
mclapply() function. If your computer supports multi-core computation, then using 
hoo.mc.ena.accumulate.data() could speed up your computation.
}
\examples{
accum = hoo.mc.ena.accumulate.data(data = mock,
                                   Units = c("site", "userName"),
                                   Conversation = c("site"),
                                   Codes = c("Code1", "Code2", "Code3", "Code4"),
                                   dataModeCol = "data",
                                   modeObserve = "chat",
                                   usersCol = "userName",
                                   windowSize = 4)

}
