% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoo.mc.horizon.R
\name{hoo.mc.horizon}
\alias{hoo.mc.horizon}
\title{Horizon of Observation Analysis for Multimodal Data}
\usage{
hoo.mc.horizon(data, Units, Conversation, Codes, dataModeCol, modeObserve,
  usersCol, windowSize)
}
\arguments{
\item{data}{Multimodal data.frame or data.table}

\item{Units}{A vector of Strings describing the Units for ENA model}

\item{Conversation}{A vector of Strings describing the Conversations took place in ENA model}

\item{Codes}{A vector of Strings entailing the big C Codes of interest for ENA
analysis}

\item{dataModeCol}{Name of the column where the types of multimodal data is stored}

\item{modeObserve}{Modes of data where actions are observable to all players}

\item{usersCol}{Name of the column entailing the unique user tracking info}

\item{windowSize}{Size of the moving stanza window, for looking backwards (for whole
conversation, input 1)}
}
\value{
a data frame containing the adjacency vectors of each ENA Units within data
}
\description{
For multimodal data, players might not fully observe other player's actions. This function takes
such factor into account and return the adjacency matrix for the actions players can observe.
}
\details{
Specifically, this function utilizes parallel package's mclapply() function. If your computer 
supports multi-core computation, then using hoo.mc.horizon() could speed up your computation.
}
\examples{
adj = hoo.mc.horizon(data = mock,
                     Units = c("site", "userName"),
                     Conversation = c("site"),
                     Codes = c("Code1", "Code2", "Code3", "Code4"),
                     dataModeCol = "data",
                     modeObserve = "chat",
                     usersCol = "userName",
                     windowSize = 4)

}
