% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoo.mock.R
\docType{data}
\name{mock}
\alias{mock}
\title{Mock data to test horizon function}
\format{An object of class \code{data.frame} with 18 rows and 7 columns.}
\usage{
mock
}
\description{
Mock data to test horizon function
}
\keyword{datasets}
