% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoo.horizon.R
\name{hoo.horizon}
\alias{hoo.horizon}
\title{Horizon of Observation Analysis for Multimodal Data}
\usage{
hoo.horizon(data, Units, Conversation, Codes, dataModeCol, modeObserve,
  usersCol, windowSize)
}
\arguments{
\item{data}{The original data containing the multimodal data (input a data frame or a data table)}

\item{Units}{ENA Units used on the original data (need to input a character string or a vector of character strings)}

\item{Conversation}{ENA Conversations within the original data (need to input a character string or a vector of character strings)}

\item{Codes}{Big C Codes used in the original data (need to input a character string or a vector of character strings)}

\item{dataModeCol}{Name of the column where the types of multimodal data is stored (need to input a character string)}

\item{modeObserve}{Name of the mode of data where other players can observe their actions (need to input a character string or a vector of character strings)}

\item{usersCol}{Name of the column where usernames or userid (something to distinguish players in the game) is stored (need to input a character string)}

\item{windowSize}{Size of the moving stanza window, for looking backwards (need to input a numeric value; for whole conversation, input 1)}
}
\value{
a data frame containing the adjacency vectors of each ENA Units within your data
}
\description{
For multimodal data, players might not fully observe other player's actions. This function takes such factor into account and return the adjacency matrix for the actions players can observe.
}
\examples{
adj = hoo.horizon(data = mock,
                  Units = c("site", "userName"),
                  Conversation = c("site"),
                  Codes = c("Code1", "Code2", "Code3", "Code4"),
                  dataModeCol = "data",
                  modeObserve = "chat",
                  usersCol = "userName",
                  windowSize = 4)

}
